style: move IkarosSizeSource to IkarosClient.kt
